update pivot_union to use binhist consumer to gen pivots
pivot_union() now puts its output into a BinHistogram<double,float>
so that we can use a binhist consumer to generate equal weight pivots
in ComputeAggregatePivots() (so resample_bins_irregular() is no longer
needed).   also update pivot_bench to use this code path.

+++ src/carp/rtp.h
@@ -91,7 +91,7 @@ class RTP {
   Status HandlePivotBroadcast(void* buf, unsigned int bufsz, int src);
 
   Status ReplayBegin();
-  void ComputeAggregatePivots(int stage_num, int num_merged,
+  void ComputeAggregatePivots(int stage_num, size_t num_merged,
                               double* merged_pivots, double& merged_weight);
 
   Carp* carp_;
